Move modularized functions to separate file
